Get total for each diagnosis
